Test automation: Added globalSetup step for all tests, refactored roles & permissions tests accordingly, updated test script
